+++ cli/tests/pcluster/config/test_section_cluster.py
@@ -38,6 +38,8 @@ from tests.pcluster.config.defaults import DefaultCfnParams, DefaultDict
                     "base_os": "alinux2",
                     "scheduler": "slurm",
                     "cluster_config_metadata": {"sections": {"cluster": ["custom_cluster_label"]}},
+                    "master_instance_type": "t2.micro",
+                    "compute_instance_type": "t2.micro",
                 },
             ),
             "custom_cluster_label",
@@ -50,6 +52,8 @@ from tests.pcluster.config.defaults import DefaultCfnParams, DefaultDict
                     "additional_iam_policies": ["arn:aws:iam::aws:policy/CloudWatchAgentServerPolicy"],
                     "base_os": "alinux2",
                     "scheduler": "slurm",
+                    "master_instance_type": "t2.micro",
+                    "compute_instance_type": "t2.micro",
                 },
             ),
             "default",
@@ -86,6 +90,8 @@ from tests.pcluster.config.defaults import DefaultCfnParams, DefaultDict
                         "arn:aws:iam::aws:policy/AWSBatchFullAccess",
                         "arn:aws:iam::aws:policy/CloudWatchAgentServerPolicy",
                     ],
+                    "master_instance_type": "t2.micro",
+                    "compute_instance_type": "t2.micro",
                 },
             ),
             "default",
@@ -267,7 +273,6 @@ def test_hit_cluster_section_from_file(mocker, config_parser_dict, expected_dict
         ("placement", "cluster", "cluster", None),
         # Head node
         # TODO add regex for master_instance_type
-        ("master_instance_type", None, "t2.micro", None),
         ("master_instance_type", "", "", None),
         ("master_instance_type", "test", "test", None),
         ("master_instance_type", "NONE", "NONE", None),
@@ -281,7 +286,6 @@ def test_hit_cluster_section_from_file(mocker, config_parser_dict, expected_dict
         ("master_root_volume_size", "31", 31, None),
         # Compute fleet
         # TODO add regex for compute_instance_type
-        ("compute_instance_type", None, "t2.micro", None),
         ("compute_instance_type", "", "", None),
         ("compute_instance_type", "test", "test", None),
         ("compute_instance_type", "NONE", "NONE", None),
@@ -543,7 +547,6 @@ def test_sit_cluster_param_from_file(
         ("shared_dir", "NONE", "NONE", None),  # NONE is evaluated as a valid path
         # Head node
         # TODO add regex for master_instance_type
-        ("master_instance_type", None, "t2.micro", None),
         ("master_instance_type", "", "", None),
         ("master_instance_type", "test", "test", None),
         ("master_instance_type", "NONE", "NONE", None),
@@ -801,6 +804,9 @@ def test_sit_cluster_section_to_file(mocker, section_dict, expected_config_parse
 def test_cluster_section_to_cfn(
     mocker, cluster_section_definition, section_dict, expected_cfn_params, default_threads_per_core
 ):
+    section_dict["master_instance_type"] = "t2.micro"
+    if cluster_section_definition == CLUSTER_SIT:
+        section_dict["compute_instance_type"] = "t2.micro"
     utils.set_default_values_for_required_cluster_section_params(section_dict)
     utils.mock_pcluster_config(mocker)
     mocker.patch("pcluster.config.cfn_param_types.get_efs_mount_target_id", return_value="valid_mount_target_id")